% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EventData.R
\docType{class}
\name{EventDate-class}
\alias{EventDate-class}
\alias{.eventDate}
\title{\code{EventDate-class}}
\description{
Stores event data and plot formatting information.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame containing the necessary data}

\item{\code{columns}}{A vector of columns contain the event}

\item{\code{label}}{A short name for the event}

\item{\code{lwd}}{Line width}

\item{\code{col}}{The colour to use for each event}
}}

