% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EventData.R
\name{EventDate}
\alias{EventDate}
\title{EventDate class}
\usage{
EventDate(data, columns, label = columns[1], lwd = 4, col = "grey")
}
\arguments{
\item{data}{A data.frame containing the necessary data}

\item{columns}{A vector of columns contain the event}

\item{label}{A short name for the event}

\item{lwd}{Line width}

\item{col}{The colour to use for each event}
}
\value{
An object of class \code{\link[=EventDate-class]{EventDate}}
}
\description{
#' A constructor for the \code{\link[=EventDate-class]{EventDate}} class.
}
