% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EventDuration.R
\name{EventDuration}
\alias{EventDuration}
\title{\code{EventDuration} class constructor}
\usage{
EventDuration(data, startDateColumns, endDateColumns,
  label = startDateColumns[1], col = "blue", ypos = c(0.2, 0.8))
}
\arguments{
\item{data}{A data.frame containing the necessary data}

\item{startDateColumns}{A vector of columns contain the start of treatment times}

\item{endDateColumns}{A vector of columns contain the end of treatment times. These are paired with startDateColumns}

\item{label}{A short name for the treatment}

\item{col}{The colour to use for each treatment}

\item{ypos}{A vector of length 2. This specifies the vertical min and max to plot the rect in}
}
\value{
An object of class \code{\link[=EventDuration-class]{EventDuration}}
}
\description{
A constructor for the \code{\link[=EventDuration-class]{EventDuration}} class.
}
