% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_AddEvent.R
\name{TreatmentPlot_AddEvent}
\alias{TreatmentPlot_AddEvent}
\title{AddEvent}
\arguments{
\item{event}{An object of class \code{\link{EventDate}}}

\item{ypos}{A vector of length 2 defining the vertical location of the line segment within a patients row. Should be between 0 and 1}
}
\description{
Adds line segments to the plot at the date specified in event. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
