% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_AddEvent.R
\name{TreatmentPlot_AddEventDuration}
\alias{TreatmentPlot_AddEventDuration}
\title{AddEventDuration}
\arguments{
\item{eventDuration}{An object of class \code{\link{EventDuration}}}

\item{ypos}{A vector of length 2 defining the vertical location of the line segment within a patients row. Should be between 0 and 1}
}
\description{
Adds rectangles to the plot between the dates specified in eventDuration. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
