% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_AddEvent.R
\name{TreatmentPlot_AddEventLabel}
\alias{TreatmentPlot_AddEventLabel}
\title{AddEvent}
\arguments{
\item{event}{An object of class \code{\link{EventDate}}}

\item{cex}{Relative font size}

\item{pos}{A position specifier for the text. If specified this overrides any adj value given. Values of 1, 2, 3 and 4, respectively indicate positions below, to the left of, above and to the right of the specified coordinates.}
}
\description{
Adds labels to events. This acts as a wrapper of \code{\link[graphics]{text}}. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
