% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_Add.R
\name{TreatmentPlot_AddGrid}
\alias{TreatmentPlot_AddGrid}
\title{AddGrid}
\arguments{
\item{xDistance}{Distance between major grid lines in days in the x direction}

\item{xSubDistance}{Distance between sub grid lines in days in the x direction. These are dashed}

\item{col}{the colour of the gridlines}

\item{lwd}{the line weight of the gridlines}
}
\description{
Adds gridlines to the plot. All parameters are optional. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
