% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_Add.R
\name{TreatmentPlot_AddLegend}
\alias{TreatmentPlot_AddLegend}
\title{AddLegend}
\arguments{
\item{itemsOnRow}{The number of items to display on each row}

\item{mar}{Give the legend a margin (is passed to par). The default is c(0,0,0,0). Each entry corresponds to bottom, left, top, right}

\item{xoffset}{An offset parameter in the x direction. Should be between 0 and 1}
}
\description{
Adds the legend to the plot. Items to add to the legend are added to the TreatmentPlot object automatically. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
