% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_Add.R
\name{TreatmentPlot_AddPatientKey}
\alias{TreatmentPlot_AddPatientKey}
\title{AddPatientKey}
\arguments{
\item{xpos}{The position on the xaxis. Default is 0, these will be left-aligned to the right of this point}

\item{col}{The colour of the text}

\item{cex}{The size of the text}
}
\description{
Adds the patientKeys to the plot. All parameters are optional. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
