% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_Add.R
\name{TreatmentPlot_AddTreatmentYAxis}
\alias{TreatmentPlot_AddTreatmentYAxis}
\title{AddTreatmentYAxis}
\arguments{
\item{treatmentLabels}{A character vector of treatment labels to use instead of the codes stored in the \code{\link{TreatmentPlot}} object.}

\item{xSubDistance}{Distance between sub grid lines in days in the x direction. These are dashed}

\item{lwd}{The line weight of the axis}

\item{cex}{The text size}
}
\description{
Adds treatment labels to the yaxis. All parameters are optional. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
