% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_Add.R
\name{TreatmentPlot_AddXAxis}
\alias{TreatmentPlot_AddXAxis}
\title{AddXAxis}
\arguments{
\item{every}{Distance between each tick. On the scale of scale}

\item{scale}{A character string one of "days","weeks","months","years". Default is days}

\item{lwd}{The line weight of the axis}
}
\description{
Adds the xasis to the plot. All parameters are optional. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
