% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_Add.R
\name{TreatmentPlot_AddXLabel}
\alias{TreatmentPlot_AddXLabel}
\title{AddXLabel}
\arguments{
\item{xlab}{The label}

\item{cex}{The size of the text}

\item{line}{The distance to draw the label from the xaxis}
}
\description{
Adds a label to the xaxis. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
