% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot.R
\name{TreatmentPlot_CloseScreens}
\alias{TreatmentPlot_CloseScreens}
\title{CloseScreens}
\description{
Closes the screens created by \code{\link[=TreatmentPlot_CreatePlot]{CreatePlot}}.  This is a method of the reference class \code{\link{TreatmentPlot}}.
}
\details{
This class uses two screens. The reference to each is stored in \code{object$screens}. \code{object$screens[1]} is the screen number for the main plot area. \code{object$screens[2]} is te screen number for the legend. To add to screen i use \code{screen(i)}. For more details see \code{\link[graphics]{screen}}.
}
