% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPlot_CreatePlot.R
\name{TreatmentPlot_CreatePlot}
\alias{TreatmentPlot_CreatePlot}
\title{CreatePlot}
\arguments{
\item{xlim}{The range of values in the x direction}

\item{numberRowsLegend}{Is used to define the space for the legend screen size. If zero then there will be no space for the legend and you should avoid calling AddLegend}

\item{xlab}{Add a xaxis title. Consider using AddXLabel for more flexibility}

\item{ylab}{Add a yaxis title. Not recommended}

\item{mar}{set the mar parameter of par for the main plot area}
}
\description{
Prepares the plot regions. Must be called prior to calling other plot functions. This is a method of the reference class \code{\link{TreatmentPlot}}.
}
